types破壊したいが気力が足りなかった
