+++ src/aggregator/aggregator.ts
@@ -1,7 +1,8 @@
 import { getCanvasSize } from "@/camera/camera";
+import { Resolution } from "@/rendering/rendering";
 import { Rect } from "../math/rect";
 import { bufferLocalLogicalIndex } from "../rendering/rendering";
-import { IterationBuffer, Resolution } from "../types";
+import { IterationBuffer } from "../types";
 
 // FIXME: 配列全部舐める必要があるのよくないので良い感じにデータを持つようにする
 let iterationCache: IterationBuffer[] = [];
